requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
grpcio==1.66.1
grpcio-tools==1.66.1
idna==3.10
protobuf==5.28.2
requests==2.32.3
urllib3==2.2.3
